requirements.txt
numpy==1.19.5
opencv-python==4.5.3
PyQt5==5.15.4
Pillow==8.2.0
